backend/requirements.txt
fastapi
uvicorn
python-dotenv
byllm==0.4.5
requests
redis
pydantic
fakeredis
httpx
pydantic-settings
jaseci==2.0.0
jaclang==0.8.10
python-multipart
